Fix wifi7 example remote station manager

+++ scratch/wifi7-industrial-mlo.cc
@@ -52,8 +52,11 @@ int main(int argc, char* argv[])
     // Wi-Fi documentation notes that EMLSR and some MLO features are
     // still experimental and can cause crashes; this example avoids
     // enabling them explicitly.
-    wifi.SetRemoteStationManager(0, "ns3::MinstrelHtWifiManager");
-    wifi.SetRemoteStationManager(1, "ns3::MinstrelHtWifiManager");
+    wifi.SetRemoteStationManager(static_cast<uint8_t>(0), "ns3::MinstrelHtWifiManager");
+    wifi.SetRemoteStationManager(static_cast<uint8_t>(1), "ns3::MinstrelHtWifiManager");
+
+    //wifi.SetRemoteStationManager(0, "ns3::MinstrelHtWifiManager");
+    //wifi.SetRemoteStationManager(1, "ns3::MinstrelHtWifiManager");
 
     SpectrumWifiPhyHelper phy(2);
     phy.SetPcapDataLinkType(WifiPhyHelper::DLT_IEEE802_11_RADIO);
